+++ CHANGES.rst
@@ -4,7 +4,8 @@ CHANGELOG
 unreleased
 -------
 
-- [breaking] Add support for MySQL 5.7.6 and up. Drop support for older MySQL versions.
+- Add support for MySQL 5.7.6 and up with new configuration options.
+    Legacy configuration supports older MySQL and MariaDB
 - [breaking] mysql_exec ini option replaced with mysql_mysqld_safe
 - [breaking] --mysql-exec cmd option replaced with --mysql-mysqld-safe
 - [breaking] dropped mysql_init ini option, --mysql-init cmd option